Move MergeDups goal to src/goals/MergeDupGoal
